backend/requirements.txt
anthropic
cerebras_cloud_sdk
fastapi
langchain-text-splitters
openai
pgvector
psycopg2-binary
PyJWT
pydantic
pydantic[email]
python-dotenv
python-multipart
requests
sqlalchemy
uvicorn

# # to enable use in jupyter lab 
# ipykernel


